backend/requirements.txt
django==5.2.3
django-cors-headers==4.7.0
django-ninja==1.4.3 